# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IES Backend",
    "description": "Intelligent Exploration System - Entity extraction and knowledge management",
    "version": "0.1.0"
  },
  "paths": {
...
    "/journeys": {
      "post": {
        "tags": [
          "journeys",
          "journeys"
        ],
        "summary": "Create Journey",
        "description": "Create a new exploration journey.\n\nThis endpoint saves a breadcrumb journey from Readest or other exploration\ninterfaces. Journeys capture the path, marks, and thinking partner exchanges\nduring exploration.",
        "operationId": "create_journey_journeys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/journeys/{journey_id}": {
      "get": {
        "tags": [
          "journeys",
          "journeys"
        ],
        "summary": "Get Journey",
        "description": "Get a journey by ID.\n\nReturns the complete journey including path, marks, and thinking partner\nexchanges.",
        "operationId": "get_journey_journeys__journey_id__get",
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreadcrumbJourney"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "journeys",
          "journeys"
        ],
        "summary": "Update Journey",
        "description": "Update a journey.\n\nUse this to:\n- Append new steps to the path (continuing exploration)\n- Add marks (highlights, annotations)\n- Add thinking partner exchanges\n- Set end time when exploration completes\n- Update title, tags, or notes",
        "operationId": "update_journey_journeys__journey_id__patch",
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreadcrumbJourney"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "journeys",
          "journeys"
        ],
        "summary": "Delete Journey",
        "description": "Delete a journey.\n\nRemoves the journey and all associated steps, marks, and exchanges.",
        "operationId": "delete_journey_journeys__journey_id__delete",
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Journey Journeys  Journey Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/journeys/user/{user_id}": {
      "get": {
        "tags": [
          "journeys",
          "journeys"
        ],
        "summary": "List User Journeys",
        "description": "List journeys for a user.\n\nReturns paginated list of journeys ordered by start time (newest first).",
        "operationId": "list_user_journeys_journeys_user__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BreadcrumbJourney": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "entry_point": {
            "$ref": "#/components/schemas/EntryPoint"
          },
          "path": {
            "items": {
              "$ref": "#/components/schemas/JourneyStep"
            },
            "type": "array",
            "title": "Path"
          },
          "marks": {
            "items": {
              "$ref": "#/components/schemas/JourneyMark"
            },
            "type": "array",
            "title": "Marks"
          },
          "thinking_partner_exchanges": {
            "items": {
              "$ref": "#/components/schemas/ThinkingPartnerExchange"
            },
            "type": "array",
            "title": "Thinking Partner Exchanges"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "processed": {
            "type": "boolean",
            "title": "Processed",
            "default": false
          },
          "siyuan_note_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Siyuan Note Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "entry_point"
        ],
        "title": "BreadcrumbJourney",
        "description": "Complete record of an exploration journey.",
        "example": {
          "entry_point": {
            "reference": "acceptance",
            "type": "search"
          },
          "marks": [
            {
              "content": "Acceptance requires active engagement with reality",
              "entity_id": "concept_123",
              "type": "highlight"
            }
          ],
          "path": [
            {
              "dwell_time_seconds": 45.2,
              "entity_id": "concept_123",
              "entity_name": "Acceptance"
            },
            {
              "dwell_time_seconds": 30.0,
              "entity_id": "concept_456",
              "entity_name": "Resignation"
            }
          ],
          "user_id": "chris"
        }
      },
...
      "EntryPoint": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/EntryPointType"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "type",
          "reference"
        ],
        "title": "EntryPoint",
        "description": "How a journey was initiated."
      },
      "EntryPointType": {
        "type": "string",
        "enum": [
          "book",
          "search",
          "dashboard",
          "entity",
          "external"
        ],
        "title": "EntryPointType",
        "description": "How a journey was started."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JourneyCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "entry_point": {
            "$ref": "#/components/schemas/EntryPoint"
          },
          "path": {
            "items": {
              "$ref": "#/components/schemas/JourneyStep"
            },
            "type": "array",
            "title": "Path"
          },
          "marks": {
            "items": {
              "$ref": "#/components/schemas/JourneyMark"
            },
            "type": "array",
            "title": "Marks"
          },
          "thinking_partner_exchanges": {
            "items": {
              "$ref": "#/components/schemas/ThinkingPartnerExchange"
            },
            "type": "array",
            "title": "Thinking Partner Exchanges"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "entry_point"
        ],
        "title": "JourneyCreateRequest",
        "description": "Request to create/save a journey."
      },
      "JourneyCreateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "siyuan_note_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Siyuan Note Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Journey saved successfully"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "JourneyCreateResponse",
        "description": "Response after creating a journey."
      },
      "JourneyListResponse": {
        "properties": {
          "journeys": {
            "items": {
              "$ref": "#/components/schemas/BreadcrumbJourney"
            },
            "type": "array",
            "title": "Journeys"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "default": 20
          }
        },
        "type": "object",
        "required": [
          "journeys",
          "total"
        ],
        "title": "JourneyListResponse",
        "description": "Response for listing journeys."
      },
      "JourneyMark": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/MarkType"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "source_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Location"
          }
        },
        "type": "object",
        "required": [
          "type",
          "entity_id",
          "content"
        ],
        "title": "JourneyMark",
        "description": "A mark (highlight, annotation, etc.) made during exploration."
      },
      "JourneyStep": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "entity_name": {
            "type": "string",
            "title": "Entity Name"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "source_passage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Passage"
          },
          "source_book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Book Id"
          },
          "dwell_time_seconds": {
            "type": "number",
            "title": "Dwell Time Seconds",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "entity_name"
        ],
        "title": "JourneyStep",
        "description": "A single step in an exploration journey."
      },
      "JourneyUpdateRequest": {
        "properties": {
          "path": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/JourneyStep"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "marks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/JourneyMark"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marks"
          },
          "thinking_partner_exchanges": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ThinkingPartnerExchange"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking Partner Exchanges"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "JourneyUpdateRequest",
        "description": "Request to update a journey (add steps, marks, etc.)."
      },
...
      "MarkType": {
        "type": "string",
        "enum": [
          "highlight",
          "annotation",
          "question",
          "bookmark"
        ],
        "title": "MarkType",
        "description": "Types of marks made during exploration."
      },
...
      "ThinkingPartnerExchange": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "entity_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Context"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "ThinkingPartnerExchange",
        "description": "A question/response exchange with the thinking partner."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}